+++ NEWS.md
@@ -5,6 +5,10 @@
 * Remove unnecessary imports.
 * Revised `lbl_df()` due to changes in the internals of `tibble::trunc_mat()`.
 
+## New functions
+
+* `as_factor()` to convert labelled vectors into factors, preserving labels.
+
 ## Changes to functions
 
 * `get_dv_labels()` now supports `brmsfit`-objects from package `brms`.